// core/domain/enums/dividend.enum.ts
export enum DividendFrequencyEnum {
    MONTHLY = "monthly",
    QUARTER = "quarter",
    ANNUAL = "annual",
    SEMI_ANNUAL = "semi_anl",
    THREE_TIMES_A_YEAR = "3x_a_yr",
    IRREGULAR = "irreg",
    SUSPENDED = "suspend",
    OTHER = "other",
    NONE = "none"
}

export enum DividendTypeEnum {
    REGULAR_CASH = "regular_cash",
    SPECIAL_CASH = "special_cash",
    RETURN_OF_CAPITAL = "return_of_capital",
    RETURN_OF_PREMIUM = "return_prem.",
    FINAL = "final",
    INTERIM = "interim",
    SECOND_INTERIM = "2nd_interim",
    DISCONTINUED = "discontinued",
    CANCELLED = "cancelled",
    OMITTED = "omitted",
    OTHER = "other",
}